Backend Szkielet v3

+++ Backend/src/main/java/com/PZ/AnkietBud/repositories/SurveyRepository.java
@@ -1,15 +1,22 @@
 package com.PZ.AnkietBud.repositories;
 
-import com.PZ.AnkietBud.TMP.Customer;
-import com.PZ.AnkietBud.TMP.Customer2;
 import com.PZ.AnkietBud.classes.Survey;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface SurveyRepository extends CrudRepository<Survey, Long> {
     List<Survey> findAll();
-    Survey findById(long id);
+//    Optional<Survey> findById(Integer id);
+    Survey findById(Integer id);
+
+//    @Query("select * from survey where name=:name")
+//    List<Survey> findByName(@Param("name") String name);
+//
+//    @Modifying
+//    @Query("UPDATE person SET first_name = :name WHERE id = :id")
+//    boolean updateByFirstName(@Param("id") Long id, @Param("name") String name);
 }
